+++ src/stp/invariants/HaltGameStateInvariant.cpp
@@ -6,7 +6,7 @@
 #include <utilities/GameStateManager.hpp>
 
 namespace rtt::ai::stp::invariant {
-double HaltGameStateInvariant::metricCheck(world_new::view::WorldDataView world, const Field *field) const noexcept {
+uint8_t HaltGameStateInvariant::metricCheck(world_new::view::WorldDataView world, const Field *field) const noexcept {
     return GameStateManager::getCurrentGameState().getRuleSet().title == "halt" ? 1.0 : 0.0;
 }
 }  // namespace rtt::ai::stp::invariant
