+++ src/stp/invariants/WeHaveBallInvariant.cpp
@@ -5,7 +5,7 @@
 #include "stp/invariants/WeHaveBallInvariant.h"
 
 namespace rtt::ai::stp::invariant {
-double WeHaveBallInvariant::metricCheck(world_new::view::WorldDataView world, const Field *field) const noexcept {
+uint8_t WeHaveBallInvariant::metricCheck(world_new::view::WorldDataView world, const Field *field) const noexcept {
     auto& us = world.getUs();
     return std::any_of(us.begin(), us.end(), [](auto& robot) { return robot.hasBall(); }) ? 1.0 : 0.0;
 }
